requirements.txt
fastapi
uvicorn
python-dotenv
stripe
requests
pydantic

# ── New for authentication ──
pyjwt[crypto]          # JWT encoding/decoding
passlib[argon2]